requirements.txt
# Core Data Analysis
pandas
numpy
yfinance

# Advanced Visualization
matplotlib
seaborn
plotly
mplfinance
bokeh

# Technical Analysis
ta-lib
ta

# Machine Learning
scikit-learn
xgboost

# Statistical Analysis
scipy
statsmodels

# Data Storage & Processing
sqlalchemy
sqlite3
h5py

# Web Framework for Dashboard
dash
dash-bootstrap-components
flask

# Utilities
python-dotenv
pyyaml
tqdm
colorama

# Testing
pytest
pytest-cov

# Documentation
sphinx
sphinx-rtd-theme
